+++ GaluhadPlugins/Pets/Strings/English/Pets.lua
@@ -406,6 +406,8 @@ _PETSTRINGS[327] = {[1]="Ill Omened Beast";[2]="Available from the Skirmishes --
 _PETSTRINGS[328] = {[1]="Ill Omened Crawler";[2]="Available from the Skirmishes -- Ill Omens event."}
 _PETSTRINGS[329] = {[1]="Snow-covered Aurochs";[2]="Limited time availability: Legendary Item Reward Track Season 5."}
 _PETSTRINGS[330] = {[1]="Fiery Blue Beetle";[2]="Limited time availability: Legendary Item Reward Track Season 5."}
+_PETSTRINGS[331] = {[1]="White Puppy";[2]="Available through barter during the Spring Festival."}
+_PETSTRINGS[332] = {[1]="Adorable Sickle-fly";[2]="Available through barter during the Spring Festival."}
 
 _PETSTRINGS[500] = {[1]="Stripey Bat";[2]=""}
 _PETSTRINGS[501] = {[1]="Injured Bat";[2]=""}
@@ -425,9 +427,8 @@ _PETSTRINGS[514] = {[1]="Winter's Wind";[2]="";}
 _PETSTRINGS[515] = {[1]="Frost-tusk Youngling";[2]="";}
 _PETSTRINGS[516] = {[1]="Ferocious Battle-kitten";[2]="";}
 _PETSTRINGS[517] = {[1]="Snow Seagull";[2]=""}
-_PETSTRINGS[518] = {[1]="Adorable Sickle-fly";[2]=""}
-_PETSTRINGS[519] = {[1]="Rohan Aurochs";[2]=""}
-_PETSTRINGS[520] = {[1]="Twilight Fire-fly";[2]=""}
-_PETSTRINGS[521] = {[1]="Deep-claw Whelp";[2]=""}
-_PETSTRINGS[522] = {[1]="Amber Fire-fly";[2]=""}
-_PETSTRINGS[523] = {[1]="Evergreen Sickle-fly";[2]=""}
+_PETSTRINGS[518] = {[1]="Rohan Aurochs";[2]=""}
+_PETSTRINGS[519] = {[1]="Twilight Fire-fly";[2]=""}
+_PETSTRINGS[520] = {[1]="Deep-claw Whelp";[2]=""}
+_PETSTRINGS[521] = {[1]="Amber Fire-fly";[2]=""}
+_PETSTRINGS[522] = {[1]="Evergreen Sickle-fly";[2]=""}