+++ Changelog.md
@@ -1,4 +1,10 @@
 Change log:
+* 2023-03-10 - U34.3 - 1 new + 1 newly available
+  * White Puppy, Adorable Sickle-fly
+* 2023-02-02 - U34.2 - 2 new + 2 newly available + 9 future pets
+  * newly available: Ill Omened Beast, Ill Omened Crawler
+  * new from reward track: Snow-covered Aurochs, Fiery Blue Beetle
+  * new feature: Frost-tusk Youngling, Ferocious Battle-kitten, Snow Seagull, Adorable Sickle-fly, Rohan Aurochs, Twilight Fire-fly, Deep-claw Whelp, Amber Fire-fly, Evergreen Sickle-fly
 * 2022-12-15 - U34.1.1 Rakish Feline, Dashing Feline, Charming Feline
 * 2022-12-08 - U34.1 Festive Yule Goat Kid is now available
 * 2022-11-16 - U34